Improved selection to fit Starcraft 1 profile
Now, clicking or dragging over an empty space does nothing; but,
dragging over or clicking on another unit selects a different group. If
shift (the "append" action) is held down, the new units are added to the
selected group.

The only drawback now is that Starcraft 1 only allowed eight units to be
selected at a time, which should be resolved in this tomorrow.