Got behavior buttons working
Method is reasonably flexible and allows for advanced behaviors.
ControlButton handles the button's appearance and action (see about
dividing it per MVC). It sets the buttons to have specific actions,
which change the Mouse. Mouse has MouseBehaviors which handle the
interaction modes with the mouse, including key modifier processing.

It might be a little better organized in the future, but the walk action
is working great.