Added basis for factions
One field, `status`, calculates whether the mission is achieved, failed,
or is TBD on the basis of a set of preconditions and conditions. If any
of the preconditions is false, the mission is failed. If all of the
conditions are true, the mission is a success. If neither of these cases
is true, then the mission is TBD.