Added second faction discretion on walk
Only the home faction can be issued commands. When multiple units are
selected, inclusive of enemy units, the enemy units do not respond to
the command.

TODO: Check whether this is how StarCraft handles it.

Unfortunately the secondary non-player faction still shows buttons,
which are unusable. It would be better to revert to a base blank box
when no controllable unit is selected.